ajout optimisation de l'analyse IA + reprise

+++ python/main.py
@@ -10,14 +10,24 @@ def main():
     file_path = "./filtered_tweets_engie.csv"
     result_path = "./result_csv/cleaned_data.csv"
 
-    df = load_data(file_path)
-    df = clean_dataframe(df)
-    df = generate_features(df)
+    # Vérifier si le fichier résultat existe déjà
+    if os.path.exists(result_path):
+        print(f"Le fichier {result_path} existe déjà. Chargement du fichier existant...")
+        df = pd.read_csv(result_path, sep=';', encoding='utf-8')
+    else:
+        print(f"Le fichier {result_path} n'existe pas. Création d'un nouveau fichier...")
+        df = load_data(file_path)
+        df = clean_dataframe(df)
+        df = generate_features(df)
 
-    df = calculateKPI()
-    df = analyze()
+        # Sauvegarder le DataFrame nettoyé avant d'analyser
+        df.to_csv(result_path, sep=';', index=False)
 
-    df.to_csv(result_path, sep=';', index=False)
+    # Analyser les tweets
+    analyze()
+
+    # Calculer les KPI après l'analyse
+    calculateKPI()
 
 if __name__ == "__main__":
     main()
